+++ test/cache/Gateway/test005.ksh
@@ -11,9 +11,17 @@ if [ -s "$input" -a -s "$CSQL_CONF" ]
 then
     REL_PATH=${PWD}/cache/Gateway
 fi
-export CSQL_CONFIG_FILE=$REL_PATH/csql.conf
-echo DSN=$DSN >> $CSQL_CONFIG_FILE
+rm -f /tmp/csql.conf
+cp $REL_PATH/csql.conf /tmp
+export CSQL_CONFIG_FILE=/tmp/csql.conf
+echo DSN=$DSN >>$CSQL_CONFIG_FILE
+
 isql $DSN < $REL_PATH/mysqlinputtest5.sql > /dev/null 2>&1
+if [ $? -ne 0 ]
+then
+    echo "DSN is not set for target db"
+    exit 1
+fi
 
 # edit /tmp/csql/csqltable.conf
 rm -f /tmp/csql/csqltable.conf /tmp/csql/csql.db
@@ -26,6 +34,12 @@ $CSQL_INSTALL_ROOT/bin/csqlserver -c >/dev/null 2>&1 &
 pid=$!
 sleep 5
 $CSQL_INSTALL_ROOT/bin/csql -g -s $REL_PATH/selectstar.sql
+if [ $? -ne 0 ]
+then
+    echo "Unable to start server"
+    exit 1
+fi
+
 $CSQL_INSTALL_ROOT/bin/csql -g -s $REL_PATH/complex.sql
 isql $DSN < $REL_PATH/dropall.sql > /dev/null 2>&1
 rm -f /tmp/csql/csqltable.conf /tmp/csql/csql.db
